+++ src/main/java/de/maxhenkel/easyvillagers/integration/waila/HUDHandlerConverter.java
@@ -2,46 +2,40 @@ package de.maxhenkel.easyvillagers.integration.waila;
 
 import de.maxhenkel.easyvillagers.blocks.tileentity.ConverterTileentity;
 import de.maxhenkel.easyvillagers.entity.EasyVillagerEntity;
+import mcp.mobius.waila.api.BlockAccessor;
 import mcp.mobius.waila.api.IComponentProvider;
-import mcp.mobius.waila.api.IDataAccessor;
-import mcp.mobius.waila.api.IPluginConfig;
-import net.minecraft.entity.merchant.villager.VillagerProfession;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.util.text.TranslationTextComponent;
-
-import java.util.List;
+import mcp.mobius.waila.api.ITooltip;
+import mcp.mobius.waila.api.config.IPluginConfig;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.entity.npc.VillagerProfession;
 
 public class HUDHandlerConverter implements IComponentProvider {
 
-    static final HUDHandlerConverter INSTANCE = new HUDHandlerConverter();
+    public static final HUDHandlerConverter INSTANCE = new HUDHandlerConverter();
 
     @Override
-    public void appendBody(List<ITextComponent> tooltip, IDataAccessor accessor, IPluginConfig config) {
-        if (!(accessor.getTileEntity() instanceof ConverterTileentity)) {
-            return;
-        }
-
-        ConverterTileentity converter = (ConverterTileentity) accessor.getTileEntity();
-        EasyVillagerEntity villagerEntity = converter.getVillagerEntity();
-        if (villagerEntity != null) {
-            if (converter.getTimer() >= ConverterTileentity.getZombifyTime() && converter.getTimer() < ConverterTileentity.getConvertTime()) {
-                VillagerProfession profession = villagerEntity.getVillagerData().getProfession();
-                if (profession.equals(VillagerProfession.NONE)) {
-                    tooltip.add(new TranslationTextComponent("entity.minecraft.zombie_villager").withStyle(TextFormatting.GRAY));
+    public void appendTooltip(ITooltip iTooltip, BlockAccessor blockAccessor, IPluginConfig iPluginConfig) {
+        if (blockAccessor.getBlockEntity() instanceof ConverterTileentity converter) {
+            EasyVillagerEntity villagerEntity = converter.getVillagerEntity();
+            if (villagerEntity != null) {
+                if (converter.getTimer() >= ConverterTileentity.getZombifyTime() && converter.getTimer() < ConverterTileentity.getConvertTime()) {
+                    VillagerProfession profession = villagerEntity.getVillagerData().getProfession();
+                    if (profession.equals(VillagerProfession.NONE)) {
+                        iTooltip.add(new TranslatableComponent("entity.minecraft.zombie_villager"));
+                    } else {
+                        iTooltip.add(new TranslatableComponent("tooltip.easy_villagers.zombie_villager_profession",
+                                new TranslatableComponent("entity.minecraft.zombie_villager"),
+                                villagerEntity.getAdvancedName()
+                        ));
+                    }
                 } else {
-                    tooltip.add(new TranslationTextComponent("tooltip.easy_villagers.zombie_villager_profession",
-                            new TranslationTextComponent("entity.minecraft.zombie_villager"),
-                            villagerEntity.getAdvancedName()
-                    ).withStyle(TextFormatting.GRAY));
-                }
-            } else {
-                ITextComponent villager = villagerEntity.getAdvancedName();
-                if (villager != null) {
-                    tooltip.add(villager);
+                    Component villager = villagerEntity.getAdvancedName();
+                    if (villager != null) {
+                        iTooltip.add(villager);
+                    }
                 }
             }
         }
     }
-
 }