+++ src/main/java/de/maxhenkel/easyvillagers/integration/waila/PluginEasyVillagers.java
@@ -1,39 +1,31 @@
 package de.maxhenkel.easyvillagers.integration.waila;
 
-import de.maxhenkel.easyvillagers.blocks.tileentity.*;
+import de.maxhenkel.easyvillagers.blocks.*;
 import mcp.mobius.waila.api.IRegistrar;
 import mcp.mobius.waila.api.IWailaPlugin;
 import mcp.mobius.waila.api.TooltipPosition;
 import mcp.mobius.waila.api.WailaPlugin;
-import net.minecraft.util.ResourceLocation;
 
 @WailaPlugin
 public class PluginEasyVillagers implements IWailaPlugin {
 
-    static final ResourceLocation OBJECT_NAME_TAG = new ResourceLocation("waila", "object_name");
-    static final ResourceLocation CONFIG_SHOW_REGISTRY = new ResourceLocation("waila", "show_registry");
-    static final ResourceLocation REGISTRY_NAME_TAG = new ResourceLocation("waila", "registry_name");
-
     @Override
     public void register(IRegistrar registrar) {
-        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, TraderTileentity.class);
-        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, AutoTraderTileentity.class);
-        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, IronFarmTileentity.class);
-        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, FarmerTileentity.class);
-        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, IncubatorTileentity.class);
+        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, TraderBlock.class);
+        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, AutoTraderBlock.class);
+        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, IronFarmBlock.class);
+        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, FarmerBlock.class);
+        registrar.registerComponentProvider(HUDHandlerVillager.INSTANCE, TooltipPosition.BODY, IncubatorBlock.class);
+        registrar.registerComponentProvider(HUDHandlerConverter.INSTANCE, TooltipPosition.BODY, ConverterBlock.class);
+        registrar.registerComponentProvider(HUDHandlerBreeder.INSTANCE, TooltipPosition.BODY, BreederBlock.class);
 
-        registrar.registerComponentProvider(HUDHandlerConverter.INSTANCE, TooltipPosition.BODY, ConverterTileentity.class);
-
-        registrar.registerComponentProvider(HUDHandlerBreeder.INSTANCE, TooltipPosition.BODY, BreederTileentity.class);
-
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, TraderTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, AutoTraderTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, IronFarmTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, FarmerTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, IncubatorTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, ConverterTileentity.class);
-        registrar.registerStackProvider(HUDHandlerVillager.INSTANCE, BreederTileentity.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, TraderBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, AutoTraderBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, IronFarmBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, FarmerBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, IncubatorBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, ConverterBlock.class);
+        registrar.registerIconProvider(HUDHandlerVillager.INSTANCE, BreederBlock.class);
     }
-
 
 }